# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This handles the game logic and communication",
        "title": "quipNotes Server",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
        "/games": {
            "post": {
                "description": "Creates a new game and returns its unique 4-digit code. Driven by the manager (host).",
                "produces": [
                    "application/json"
                ],
                "summary": "Starts a new game",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/game.CreateGameResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}": {
            "get": {
                "description": "Returns a game's code and current players. Used to validate a join.",
                "produces": [
                    "application/json"
                ],
                "summary": "Game info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.GameInfoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a game from the server. Driven by the manager (host).",
                "summary": "Ends a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/draw": {
            "post": {
                "description": "Draws Tiles (wordTiles) for a given player and a given count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Draws Tiles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "tells how many tiles to draw",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.DrawTilesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.WordsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/players": {
            "post": {
                "description": "Adds a player to a game. The playerId must be unique within the game.",
                "consumes": [
                    "application/json"
                ],
                "summary": "Joins a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "contains the player id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.AddPlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/players/{id}": {
            "delete": {
                "description": "Removes a player from a game. The playerId must exist.",
                "summary": "Leaves a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "player id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/players/{id}/tiles": {
            "get": {
                "description": "Gets all the tiles that are drawn by the player.",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets Drawn Tiles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "player id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.WordsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/submit": {
            "post": {
                "description": "Send a string array to turn in your wordTiles for the game.",
                "consumes": [
                    "application/json"
                ],
                "summary": "Turn in Note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "contains the note",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.SubmitNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{code}/submitted-notes": {
            "get": {
                "description": "Returns a list of strings that are the submitted notes for the game",
                "produces": [
                    "application/json"
                ],
                "summary": "Returns the submitted notes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Clears the submitted notes for the game",
                "summary": "Deletes the submitted notes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "game code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/game.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "game.AddPlayerRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "game.CreateGameResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "game.DrawTilesRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "game.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "game.GameInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "game.SubmitNoteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "game.WordsResponse": {
            "type": "object",
            "properties": {
                "words": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}